moved fuzzy loader from init.lua to fuzzy_loader.lua
