+++ plugin/init.lua
@@ -358,106 +358,6 @@ function pub.resurrect_on_gui_startup()
 	return suc, err
 end
 
----@alias fmt_fun fun(label: string): string
----@alias fuzzy_load_opts {title: string, description: string, fuzzy_description: string, is_fuzzy: boolean, ignore_workspaces: boolean, ignore_tabs: boolean, ignore_windows: boolean, fmt_window: fmt_fun, fmt_workspace: fmt_fun, fmt_tab: fmt_fun }
-
----Returns default fuzzy loading options
----@return fuzzy_load_opts
-function pub.get_default_fuzzy_load_opts()
-	return {
-		title = "Load State",
-		description = "Select State to Load and press Enter = accept, Esc = cancel, / = filter",
-		fuzzy_description = "Search State to Load: ",
-		is_fuzzy = true,
-		ignore_workspaces = false,
-		ignore_windows = false,
-		ignore_tabs = false,
-		fmt_workspace = function(label)
-			return wezterm.format({
-				{ Foreground = { AnsiColor = "Green" } },
-				{ Text = "󱂬 : " .. label:gsub("%.json$", "") },
-			})
-		end,
-		fmt_window = function(label)
-			return wezterm.format({
-				{ Foreground = { AnsiColor = "Yellow" } },
-				{ Text = " : " .. label:gsub("%.json$", "") },
-			})
-		end,
-		fmt_tab = function(label)
-			return wezterm.format({
-				{ Foreground = { AnsiColor = "Red" } },
-				{ Text = "󰓩 : " .. label:gsub("%.json$", "") },
-			})
-		end,
-	}
-end
-
----A fuzzy finder to restore saved state
----@param window MuxWindow
----@param pane Pane
----@param callback fun(id: string, label: string, save_state_dir: string)
----@param opts fuzzy_load_opts?
-function pub.fuzzy_load(window, pane, callback, opts)
-	wezterm.emit("resurrect.fuzzy_load.start", window, pane)
-	local state_files = {}
-
-	if opts == nil then
-		opts = pub.get_default_fuzzy_load_opts()
-	else
-		-- Merge user opts with defaults
-		local default_opts = pub.get_default_fuzzy_load_opts()
-		for k, v in pairs(default_opts) do
-			if opts[k] == nil then
-				opts[k] = v
-			end
-		end
-	end
-
-	local function insert_choices(type, fmt)
-		for _, file in ipairs(wezterm.glob("*", pub.save_state_dir .. "/" .. type)) do
-			local label
-			local id = type .. "/" .. file
-
-			if fmt then
-				label = fmt(file)
-			else
-				label = file
-			end
-			table.insert(state_files, { id = id, label = label })
-		end
-	end
-
-	if not opts.ignore_workspaces then
-		insert_choices("workspace", opts.fmt_workspace)
-	end
-
-	if not opts.ignore_windows then
-		insert_choices("window", opts.fmt_window)
-	end
-
-	if not opts.ignore_tabs then
-		insert_choices("tab", opts.fmt_tab)
-	end
-
-	window:perform_action(
-		wezterm.action.InputSelector({
-			action = wezterm.action_callback(function(_, _, id, label)
-				if id and label then
-					callback(id, label, pub.save_state_dir)
-				end
-				wezterm.emit("resurrect.fuzzy_load.finished", window, pane)
-			end),
-			title = opts.title,
-			description = opts.description,
-			fuzzy_description = opts.fuzzy_description,
-			choices = state_files,
-			fuzzy = opts.is_fuzzy,
-		}),
-		pane
-	)
-end
-
 ---@param file_path string
 function pub.delete_state(file_path)
 	wezterm.emit("resurrect.delete_state.start", file_path)
@@ -471,12 +371,10 @@ function pub.delete_state(file_path)
 end
 
 -- Export submodules
-local workspace_state = require("resurrect.workspace_state")
-pub.workspace_state = workspace_state
-local window_state = require("resurrect.window_state")
-pub.window_state = window_state
-local tab_state = require("resurrect.tab_state")
-pub.tab_state = tab_state
+pub.workspace_state = require("resurrect.workspace_state")
+pub.window_state = require("resurrect.window_state")
+pub.tab_state = require("resurrect.tab_state")
+pub.fuzzy_loader = require("resurrect.fuzzy_loader")
 
 function pub.set_max_nlines(max_nlines)
 	require("resurrect.pane_tree").max_nlines = max_nlines
